+++ src/lib/builders/index.ts
@@ -20,6 +20,7 @@ export * from './popover/index.js';
 export * from './progress/index.js';
 export * from './radio-group/index.js';
 export * from './range-calendar/index.js';
+export * from './scroll-area/index.js';
 export * from './select/index.js';
 export * from './separator/index.js';
 export * from './slider/index.js';