+++ src/lib/builders/scroll-area/events.ts
@@ -1,0 +1,8 @@
+import type { GroupedEvents } from '$lib/internal/types.js';
+
+export const scrollAreaEvents = {
+	scrollbar: ['pointerdown', 'pointerup', 'pointermove'],
+	thumb: ['pointerdown', 'pointerup'],
+} as const;
+
+export type ScrollAreaEvents = GroupedEvents<typeof scrollAreaEvents>;